fix tests and simplify cypress webpack config
