+++ tsconfig.json
@@ -28,9 +28,7 @@
     "src/docs/index.jsx",
     "eslint.config.mjs",
     "babel.config.cjs",
-    "cypress.config.mjs",
-    "webpack.config.test.mjs",
-    "webpack.common.mjs"
+    "cypress.config.mjs"
   ],
   "typeRoots": ["./node_modules/@types", "*.d.ts"]
 }
